gateway/requirements.txt
Flask==2.2.2
authlib==1.2.0
requests==2.31.0
Flask_SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
Werkzeug==2.2.2
Flask-Login==0.6.2
redis==4.5.5
#Flask-WTF==1.1.1
flask-cors==3.0.10
Flask-JWT-Extended==4.5.2